More complete fix to printf format warning than my previous attempt. CVSMAIL: CC:

+++ gasnet_trace.h
@@ -1,6 +1,6 @@
 /*   $Source: /Users/kamil/work/gasnet-cvs2/gasnet/gasnet_trace.h,v $
- *     $Date: 2005/03/12 13:03:17 $
- * $Revision: 1.39 $
+ *     $Date: 2005/04/18 05:49:17 $
+ * $Revision: 1.40 $
  * Description: GASNet Tracing Helpers (Internal code, not for client use)
  * Copyright 2002, Dan Bonachea <bonachea@cs.berkeley.edu>
  * Terms of use are as specified in license.txt
@@ -382,8 +382,8 @@ BEGIN_EXTERNC
     _GASNETI_STAT_EVENT(A,name);                                                                          \
     if (gasnet_AMGetMsgSource(token,&src) != GASNET_OK)                                                   \
       gasneti_fatalerror("gasnet_AMGetMsgSource() failed");                                               \
-    GASNETI_TRACE_PRINTF(A,(#name": src=%i handler=%i addr="GASNETI_LADDRFMT" nbytes=%i args:%s",         \
-      (int)src,(int)(handlerid),GASNETI_LADDRSTR(addr),nbytes,argstr));                                   \
+    GASNETI_TRACE_PRINTF(A,(#name": src=%i handler=%i addr="GASNETI_LADDRFMT" nbytes=%u args:%s",         \
+      (int)src,(int)(handlerid),GASNETI_LADDRSTR(addr),(unsigned int)nbytes,argstr));                     \
     GASNETI_TRACE_PRINTF(C,(#name": token: %s",                                                           \
                       gasneti_formatdata((void *)&(token), sizeof(token))));                              \
     GASNETI_TRACE_PRINTF(D,(#name": payload data: %s", gasneti_formatdata(addr,nbytes)));                 \
